address commments
